+++ .conflict-side-0/stories/0x01a.compare-target-vs-current-portfolio.md
@@ -10,6 +10,11 @@ when I do.
 
 ## Acceptance Criteria
 
+The backend `POST /portfolio/compare` API
+([#280](https://github.com/data-cartel/moneymentum/pull/280)) supplies the
+per-position target/current/delta data and the tradable flag these criteria
+render; the criteria themselves are the frontend portfolio surface.
+
 - [ ] The portfolio surface labels target and current allocations distinctly.
 - [ ] Each position row shows target weight, current weight, and the delta.
 - [ ] Positions whose delta is below the minimum tradable change are visibly
